[wpimath] Disambiguate wpimath JNI functions
Now, each collection of JNI functions has its own class.
