+++ wpimath/src/main/java/edu/wpi/first/math/Matrix.java
@@ -4,6 +4,7 @@
 
 package edu.wpi.first.math;
 
+import edu.wpi.first.math.jni.EigenJNI;
 import edu.wpi.first.math.numbers.N1;
 import java.util.Objects;
 import org.ejml.MatrixDimensionException;
@@ -360,7 +361,7 @@ public class Matrix<R extends Num, C extends Num> {
   public final <R2 extends Num, C2 extends Num> Matrix<C, C2> solveFullPivHouseholderQr(
       Matrix<R2, C2> other) {
     Matrix<C, C2> solution = new Matrix<>(new SimpleMatrix(this.getNumCols(), other.getNumCols()));
-    WPIMathJNI.solveFullPivHouseholderQr(
+    EigenJNI.solveFullPivHouseholderQr(
         this.getData(),
         this.getNumRows(),
         this.getNumCols(),
@@ -387,7 +388,7 @@ public class Matrix<R extends Num, C extends Num> {
               + this.getNumCols());
     }
     Matrix<R, C> toReturn = new Matrix<>(new SimpleMatrix(this.getNumRows(), this.getNumCols()));
-    WPIMathJNI.exp(
+    EigenJNI.exp(
         this.m_storage.getDDRM().getData(),
         this.getNumRows(),
         toReturn.m_storage.getDDRM().getData());
@@ -411,7 +412,7 @@ public class Matrix<R extends Num, C extends Num> {
               + this.getNumCols());
     }
     Matrix<R, C> toReturn = new Matrix<>(new SimpleMatrix(this.getNumRows(), this.getNumCols()));
-    WPIMathJNI.pow(
+    EigenJNI.pow(
         this.m_storage.getDDRM().getData(),
         this.getNumRows(),
         exponent,
@@ -709,7 +710,7 @@ public class Matrix<R extends Num, C extends Num> {
    * @param lowerTriangular Whether this matrix is lower triangular.
    */
   public void rankUpdate(Matrix<R, N1> v, double sigma, boolean lowerTriangular) {
-    WPIMathJNI.rankUpdate(this.getData(), this.getNumRows(), v.getData(), sigma, lowerTriangular);
+    EigenJNI.rankUpdate(this.getData(), this.getNumRows(), v.getData(), sigma, lowerTriangular);
   }
 
   @Override
