+++ wpimath/src/main/java/edu/wpi/first/math/StateSpaceUtil.java
@@ -5,6 +5,7 @@
 package edu.wpi.first.math;
 
 import edu.wpi.first.math.geometry.Pose2d;
+import edu.wpi.first.math.jni.StateSpaceUtilJNI;
 import edu.wpi.first.math.numbers.N1;
 import edu.wpi.first.math.numbers.N3;
 import edu.wpi.first.math.numbers.N4;
@@ -102,7 +103,8 @@ public final class StateSpaceUtil {
    */
   public static <States extends Num, Inputs extends Num> boolean isStabilizable(
       Matrix<States, States> A, Matrix<States, Inputs> B) {
-    return WPIMathJNI.isStabilizable(A.getNumRows(), B.getNumCols(), A.getData(), B.getData());
+    return StateSpaceUtilJNI.isStabilizable(
+        A.getNumRows(), B.getNumCols(), A.getData(), B.getData());
   }
 
   /**
@@ -120,7 +122,7 @@ public final class StateSpaceUtil {
    */
   public static <States extends Num, Outputs extends Num> boolean isDetectable(
       Matrix<States, States> A, Matrix<Outputs, States> C) {
-    return WPIMathJNI.isStabilizable(
+    return StateSpaceUtilJNI.isStabilizable(
         A.getNumRows(), C.getNumRows(), A.transpose().getData(), C.transpose().getData());
   }
