+++ wpimath/src/main/java/edu/wpi/first/math/geometry/Pose3d.java
@@ -8,10 +8,10 @@ import com.fasterxml.jackson.annotation.JsonAutoDetect;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import edu.wpi.first.math.WPIMathJNI;
 import edu.wpi.first.math.geometry.proto.Pose3dProto;
 import edu.wpi.first.math.geometry.struct.Pose3dStruct;
 import edu.wpi.first.math.interpolation.Interpolatable;
+import edu.wpi.first.math.jni.Pose3dJNI;
 import edu.wpi.first.util.protobuf.ProtobufSerializable;
 import edu.wpi.first.util.struct.StructSerializable;
 import java.util.Objects;
@@ -224,7 +224,7 @@ public class Pose3d implements Interpolatable<Pose3d>, ProtobufSerializable, Str
   public Pose3d exp(Twist3d twist) {
     var quaternion = this.getRotation().getQuaternion();
     double[] resultArray =
-        WPIMathJNI.expPose3d(
+        Pose3dJNI.exp(
             this.getX(),
             this.getY(),
             this.getZ(),
@@ -257,7 +257,7 @@ public class Pose3d implements Interpolatable<Pose3d>, ProtobufSerializable, Str
     var thisQuaternion = this.getRotation().getQuaternion();
     var endQuaternion = end.getRotation().getQuaternion();
     double[] resultArray =
-        WPIMathJNI.logPose3d(
+        Pose3dJNI.log(
             this.getX(),
             this.getY(),
             this.getZ(),
