+++ wpimath/src/main/native/cpp/jni/Exceptions.cpp
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-#include "WPIMathJNI_Exceptions.h"
+#include "Exceptions.h"
 
 #include <jni.h>
