+++ wpimath/src/main/native/cpp/jni/Exceptions.h
