+++ wpimath/src/test/java/edu/wpi/first/math/jni/ArmFeedforwardJNITest.java
@@ -2,15 +2,15 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package edu.wpi.first.math;
+package edu.wpi.first.math.jni;
 
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 
 import org.junit.jupiter.api.Test;
 
-public class WPIMathJNITest {
+public class ArmFeedforwardJNITest {
   @Test
   public void testLink() {
-    assertDoesNotThrow(WPIMathJNI::forceLoad);
+    assertDoesNotThrow(ArmFeedforwardJNI::forceLoad);
   }
 }
