+++ wpimath/src/test/java/edu/wpi/first/math/jni/Ellipse2dJNITest.java
@@ -1,0 +1,16 @@
+// Copyright (c) FIRST and other WPILib contributors.
+// Open Source Software; you can modify and/or share it under the terms of
+// the WPILib BSD license file in the root directory of this project.
+
+package edu.wpi.first.math.jni;
+
+import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
+
+import org.junit.jupiter.api.Test;
+
+public class Ellipse2dJNITest {
+  @Test
+  public void testLink() {
+    assertDoesNotThrow(Ellipse2dJNI::forceLoad);
+  }
+}
